requirements.txt
requests
selenium
python-dotenv